Remove update of /etc/hosts from bootcmd section
The bootcmd section commands are repeated each boot
which is O.K. for some but not for the /etc/hosts update the way
it was being done.

I am leaving only reporting commands and performing the /etc/hosts update
later in the runcmd section which does not run in subsequent boots.
I will delay the install of dovecot-imapd later in the
runcmd section rather than in the packages.  My memory says the automated
mail configuraton works better after the /etc/hosts update.

I removed the roundcube installation which defaults to interactive
configuration beyond what --yes handles.