Add Ubuntu-CloudInit directives to launch lam2
/var/www/aws/aws-nwo-lam2-Ubuntu-CloudInit.txt - Initialize lam2.duckdns.org

The lam2 instance was designed to be a SSH SOCKS5 Proxy server on port 443
Port 443 is normally used for HTTPS so is likely to be available even when
other ports are blocked.

The CloudInit directives are a fork of the lam1 directives with a new
host name that takes over the lam2 DDNS sites instead of the lam1 sites.
The lam1 http sites are all installed so this instance can check in as
lam1 and support all but the https site.

The ssl module of apache2 is not loaded so that the sshd can listen on
port 443, et-s Encrypt certificates are not installed and the default
info and status modules are invoked instead of the modified version that
serve authenticated users on the secure site.
The /var/www/lam (ssl site) directory is installed but the site is not
enabled.  The databeses for the secure site are not installed.