New lam2 CloudInit script using GitHub
The aws-lam2-Ubuntu-CloudInit.txt uses only Public GitHub repos and
the https protocol from https://github.com/LAMurakami/<repo>

It does load the private credentials cache and does not mount the efs drive.
Without the duckdns credentials it cannot check into the lam2.duckdns.org
domain.

An advantage of not using the efs drive is that it is not limited to the
Nortwest Oregon Region of AWS.

This started from the current version and the changes can be viewed with:

 git --no-pager diff --no-index \
 aws-nwo-lam2-Ubuntu-CloudInit.txt aws-lam2-Ubuntu-CloudInit.txt

I did not start with the lam1 CloudInit becasue it loads the secure
side (lam repo) which is not a public repo.