Use Elastic File System locate database
Run cron daily job for locate.

The locate command needs the database built to work.
A daily cron job does this.  The daily cron job uses nice and ionice
 to reduce impact on other jobs on the system.
I will run this after all the other initialization steps except the final
 etckeeper commit.

The Elastic File System mlocate database is built on the main server and
 should not be rebuilt on every launch of every instance.

plocate is not available on Linux Mint and is installed by default on Ubuntu 22.04
mlocate is not available on Ubuntu 22.04 even though installed as a dummy package.