The blinkenshell-sync-clone job should run first.
This job runs before any other backup jobs each day on ak20 and aws so that the daily directory is created by the lam or ubuntu user and any additional jobs run by that user can write to the same directory. Any jobs run by root can also use the same directory. By doing it this way I did not have to modify the programs. This job runs as the lam or ubuntu user so that I do not have to setup access to blinkenshell from the root user. The lam and ubuntu users share the public/private key that grants me access to blinkenshell.

There are multiple programs I created to be run as system cron jobs that write to this directory structure and create a new daily directory if it doesn't already exist.  All of them could be changed to accomodate permissions, I could create a separate program to fix permissons, or I can just schedule one of the jobs run as lam or ubuntu to run first.  I am limited to 100M on blinkenshell and don't update there often so the blinkenshell-sync-clone job only takes seconds to run so I just run it first.

In summary the blinkenshell-sync-clone job should run first until I update things.