Added public-ipv6, availability-zone and ami-id
Modified cloud-init.pl to list and update these new values and added
the new aws_apache2_ipv6.conf as an additional file to be updated.
Modified cloud-init.pl to use the = character instead of the : character
for the $var, $value split becasue the colon character is used in the
IPv6 address format.

Modified ec2-user-data.bash to use the = character instead of the : character
and added the public-ipv6, availability-zone and ami-id values to the report.

Added public-ipv6, availability-zone and ami-id to main page (html/index.html)
both in display and current values in the header comment.
Modified the current values in the header comment to use the = character.

The aws_apache2_ipv6.conf will have to be added in all the CloudInit.txt files
but I will start with /var/www/aws/aws-nwo-lam1-Ubuntu-CloudInit-ARM.txt

The aws page will not be displayed for new instances accessed over IPv6
until the additional apache configuration file is enabled.  The IPV6 address
will be on the main aws page and wor but will display the default
no-ssl website.