Update for IPv6 Apache2 implementation - part 3
Two of the cloud-init scripts fully succeed with IPv6

aws-nwo-lam2-Debian-ARM-CloudInit.txt # Starting IPv6 cloud-init
aws-nwo-lam2-Ubuntu-ARM-CloudInit.txt # Second OS more sites

 Most of the sites are not available as lam2

The lam2-Ubuntu-ARM cloud init installs all but mike, interiordems and
 oldinteriordems even though by default only no-ssl and aws are served up.
The lam site is installed but not implemented because the https port is used
 as an alternate ssh port.

The aws site was made the default by naming it 000-aws so that the server
 when accessed by IPv6 serves up this site.  This is primarily because
 to my knowledge you can't specify the IPv6 address for the Apache2
 ServerName and ServerAlias directives even though IPv4 addresses
 can be specified.

The same general changes were made:

1. Add create a new branch using the ${KEYNAME} value for every git clone repo
 that will have a file or files modified by the cloud-init process.
2. Delay cloud-init.pl until after all sites are installed
3. Remove /var/www/aws/aws_apache2_ipv6.conf site file configuration
4. Delay change of ownership until after git commit
5. Commit changes made by cloud-init to new ${KEYNAME} branch in local git repo

Output was added to the log to clarify the multiple commits using the
 same branch name.

For repositories that are public on github that source is the clone origin.

The use of the tar archive for blinkenshell was eliminated.
I may want to instead use the tar archive but include the git repo in it.