Update for IPv6 Apache2 implementation - part 4
aws-nwo-lam1-Ubuntu-CloudInit-ARM.txt # ARM Instance to take over lam1

The Instance to take over lam1 is my main cloud init as it launches a
 full clone of aws.lam1.us with all sites and by default takes over a
 number of the sites including arsc, sites, z and more.

The User data is limited to 16,384 bytes when launching an instance.
The lam1-Ubuntu-ARM cloud init exceeded this when it grew to 17,719 bytes
 with the addition of the IPv6 changes which included creating
 and commiting 13 branches named with the KEYNAME for the instance.

By removing the final commits to a separate bash script I was able
 to reduce the total size of the cloud-init to 15,548 bytes.

The separate script requres a parameter which is used in the commit comment.
 This will allow the same script to be used by other cloud-init files.