Apply EFS for REGION changes
Include aws s3 cp and ssh to aws when not Oregon changes to eliminate most
/mnt/efs dependancies except the ubuntu.tgz and ec2-user.tgz user resource
 files.

The ssh to aws connectivity allows the Oregon EFS to be used even when not
launching an instance in that region.

Instances with a public IPv4 address can use github and aws s3 without the IPv6
endpoint fix but install of the {new|old}.interiordems, mike, blinkenshell and
alaskademocrat sites which are not in the public github repos is now
dependant on the aws clone running.

The failure of the install of these four websites does not affect the
functionality of the main secure side of the website or it's ability to
become aws for the purpose of being the source of these repos.