Apply EFS for REGION changes
Include aws s3 cp and ssh to aws when not Oregon changes to eliminate most
/mnt/efs dependancies except the ubuntu.tgz and ec2-user.tgz user resource
 files.

This is a SSH SOCKS5 Proxy server on port 443 that can now be launched any
of the 29 AWS regions.  It runs dual stack and with a public IPv4 address
gives me access to IPv6 when I am at home or a resort, casino, hotel Wi-Fi
that uses a private IPv4 network and NAT gateway.