Add localhost entry to known_hosts
Use ed25519 host key type rather than the rsa key type used on
 Ubuntu and Debian systems.

Create git user for git@localhost early in the initialization process.
Update lam2.duckdns.org IPv6 address at end of initialization.

This is part of the
 https://ak20.lam1.us/a/LAM_Alaska_from_IPv4_to_IPv6#IPv6_only_workarounds
 that allow me to launch an aws instance without a public IPv4 address.

Use aws-efs-tar-extract.bash # Extract a gzip compressed tar archive
 instead of aws s3 cp to reduce the chance of an OoM (Out of Memory)
 failure on a nano instance.
This script was created as an IPv6 only workaround to eliminate the use of the
 aws command line utility in Ubuntu 24.04 Noble Numbat becasue it was being
 provided by a snap package that currenlty can't be installed on an instance
 without a public IPv4 address.  Although the aws command line utility is
 available on an Amazon Linux 2023 instance I have avoided it in this case
 becasue I want nano instance to be able to launch and Amazon Linux 2023
 uses more memory than it's predicessor Amazon Linux 2 and it's competitors
 that I use including Ubuntu and Debian.

This change allowed a nano instance launch to succeed when it had previously
 failed with an OoM (Out of Memory) error doing the lam.tgz install using
 aes s3.