Add aws-lam-gitlab-clone.bash to clone from GitLab
The aws-lam-gitlab-clone.bash script clones from GitLab using https.
This will work for both IPv6 and IPv4 unlike GitHub and eliminates the need for
the IPv6 only workaround for GitHub during initialization.
The script then sets the origin remote to use ssh, adds github and ak20 remotes,
and does a checkout to a new $keyName branch.

To use the github and ak20 remotes on an instance without a public IPv4 address
still requires the IPv6 only workaround for GitHub.