Use an annotated tag to get a date time stamp
The aws-efs-www-git-clone.bash is an IPv6 only workaround.  Since ak20 is
unavailable for an aws instance without a public IPv4 address so the private
copy of repositories in the efs on my main region are used.

If the new instance is in the us-west-2 Oregon region a localhost git user
is set up to access the repos on the efs.  If the new instance is in another
region the ssh to aws connection is used.

Once the repo is installed the remote is changed to use my main remote store
which is accessable through the GitHub ak20 IPv6 only workaround or directly
if the instance has a public IPv4 address.

https://lamurakami.github.io/blog/2024/06/05/Access-GitHub-com-from-an-instance-without-a-public-IPv4-address.html