Add default site for ssl
Point to duckdns domains using ssl

For the ssl version of the aws site NameVirtualHost is specified for the
443 port on all interfaces to allow this to work.

With name-based virtual hosting, the server relies on the client to report
the hostname as part of the HTTP headers.

IP-based resolution takes place before name-based virtual host resolution
Using a wildcard (*) for the IP address in all of the VirtualHost directives
makes IP-based mapping irrelevant and makes name-based virtual hosting easier
to implement.

On Debian and Ubuntu use:
cat /var/www/aws/ssl-NameVirtualHost.conf >> /etc/apache2/mods-available/ssl.conf

The original apache used IP-based mapping and name-based virtual hosting
was added first only for non ssl sites.  More than a decade later name-based
virtual hosting was added to the ssl module.  In both cases the client must
report the hostname as part of the HTTP headers.

This is my first step in enabling all my sites to be accessed over a secure
connection.

I made this a To Do when I noted that some browsers may default to a secure
connection.