Implement secure access to aws site
For the launch of Debian and Amazon Linux 2023 instances include initialization
of name-based virtual hosting with separate certificates for the aws and lam
sites.

The appropriate branches of the lam repo were modified to support name-based
virtual hosting on the secure side of my site.  The lam repo was the source of
the only site serverd over ssl.  The RewriteRule forcing a single domain name
for ssl was removed.  The MediaWiki LocalSettings.php contains a site name and
will still enforce a single domain name for access to the MediaWiki site.