Use Update-DuckDNS.bash for all -No-Public-IPv4
This makes the Update-DuckDNS-IPv6.bash script obsolete.

The non -No-Public-IPv4 CloudInit files may be deemed obsolete in the future.
They may use GitHub instead of GitLab as the origin for public repositories.

I did this so the new aws-lam2-openSUSE-No-Public-IPv4-CloudInit.txt could
be modified to handle an instance with a Puplic IPv4 address and I wouldn't
have to create yet another CloudInit file.