Create program to mount second EFS
I am creating a second EFS that takes advantange of Lifecycle Management
When I first created an EFS in 2017 A lot of options that are now available
were not available.  When I created the first EFS in other regions I made
it like the one in us-west-2 Oregon.

A second EFS will be created in every region and mounted as part of
initialization.

/mnt/efs # files stay in standard storage
/mnt/efs2 # files transition automatically between standard IA and archive