Updated for Debian 13
Install php-xml and libgtk2.0-0t64 in the packages section instead of the
more generic php-xml* and libgtk2.0-0 that was used in Debian 12.
Install texlive after swap has been configured to avoid a Out of Memory
error that caused initialization to fail when it was in the packages section.

Used America/Anchorage timezone because US/Alaska is not in the image.