Remove rslink file and link to rslink.bash program
I toyed with using rslink files and a program for configuration a few years ago
but never really got something easily managable and complete.

For this repo the
aws-lam-apache2-sites-configuration-{Amazon-Linux,Debian-Ubuntu,OpenSUSE}.bash
programs now create the symbolic link which may differ based on the OS.

The *CloudInit.txt files also do some of the symbolic link creation based on OS

The symbolic links for the clone and backup files are currently manually
created when an instance is made the main aws instance.